extracted socket.io logic from server file
