+++ server.js
@@ -2,79 +2,12 @@ const PORT = 8000;
 const express = require('express');
 const app = express();
 const server = require('http').Server(app);
-const io = require('socket.io')(server);
+const io = require('./src/io');
 
-const faker = require('faker');
+io.listen(server);
 
 app.use(express.static('public-react/build'));
 
-const members = {};
-
-
-io.on('connection', socket => {
-  let username = `${faker.internet.userName()}`;
-  socket.emit(
-    'message-all', 
-    {
-      user: 'admin',
-      text: members.length === 0 ?
-        `Hello ${username}, you\'re the first one here!`:
-        `Hello ${username}, you joined the chat with ${require('./utils/nameList')(Object.keys(members))}`,
-      timestamp: new Date()
-    }
-  )
-  members[username] = socket.id;
-
-  sendUserUpdate(username);
-  
-  console.log('client connected!');
-  console.log(members);
-  
-  socket.on('reset-username', newName => {
-    const oldUsername = username;
-    delete members[username];
-    username = newName;
-
-    sendUserUpdate(username, oldUsername);
-
-    members[username] = socket.id;
-  });
-
-  socket.on('message-all', msg => {
-    console.log(msg);
-    socket.broadcast.emit('message-all', msg);
-  });
-  
-  socket.on('disconnect', () => {
-    console.log('client disconnected');
-    socket.broadcast.emit(
-      'message-all', 
-      {
-        user: 'admin',
-        text: `${username} has disconnected`,
-        timestamp: new Date()
-      }
-    );
-    delete members[username];
-  });
-  
-  function sendUserUpdate(newUsername, oldUsername = null) {
-    socket.emit('set-username', { newUsername });
-    socket.broadcast.emit('member-update', { 
-      newUsername, oldUsername 
-    });
-    socket.broadcast.emit(
-      'message-all', 
-      {
-        user: 'admin',
-        text: oldUsername ?
-          `${newUsername} name changed from ${oldUsername}` :
-          `${newUsername} has joined the chat`,
-        timestamp: new Date()
-      }
-    );
-  }
-});
